requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
Django==4.1.7
django-bootstrap5==22.2
django-environ==0.9.0
django-filter==22.1
django-tz-detect==0.4.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-gis==1.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
idna==3.4
install==1.3.5
numpy==1.24.2
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7.1
ratelim==0.1.6
requests==2.28.2
routingpy==1.1.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
